bug fixing

+++ app/src/main/java/com/example/mozzart_test/data/model/PrizeCategory.kt
@@ -2,11 +2,11 @@ package com.example.mozzart_test.data.model
 
 data class PrizeCategory(
     val categoryType: Int,
-    val distributed: Int,
-    val divident: Int,
+    val distributed: Double,
+    val divident: Double,
     val fixed: Double,
     val gameType: String,
     val id: Int,
-    val jackpot: Int,
+    val jackpot: Double,
     val winners: Int
 )
